Fix handling invalid orientation in AutoRotateFilterLoader
AutoRotateFilterLoader tried to throw an exception that didn't exist
when it had received an image with invalid EXIF orientation tag.

It also used to fall back to exif_read_data when an old Imagine was used
and it wasn't tested properly which caused breaking a build.

The test is somewhat tricky. I've added an asset with a tiny JPEG that
has an EXIF orientation tag. When an old imagine is used, test reads an
image into memory, resets the orientation byte and makes ImageInterface
mock respond with the modified image when raw JPEG is requested.