Add visibility argument to flysystem resolver
Before this, there was no way to modify the visibility in a central place,
which was fine for local resolvers, or with predefined acls.

This changeset introduces a new `visibility` argument on the FlysystemResolver
that will be taken into account everytime a cache object is written to a flysystem
filesystem.

The default is public, so that cached objects are accessible from the outside,
which should be sensible.