Create filtered image in getUrlOf... methods in FilterService
By coupling the creation of the filtered image to resolvig the URL
of the resolving image the use case of requesting the URL of a filtered
image that has not been created yet (temporal coupling issue) is no
longer possible. The current implementation breaks CQS, but this
is not as bad as the previous temporal coupling.