Update FilterTrait.php
Potentially fixes #936 and #328.
Please note I am not familiar enough with the bundle to completely grasp the implications. Please double check this PR before merging :)
The previous implementation was stripping the server name from the path, preventing any remote loading to occur. If this is the right approach, happy to enhance the doc on how to make it work for remote files (just standard http)